requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
filelock==3.16.1
Flask==3.0.3
fsspec==2024.9.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
openai-whisper==20231117
python-dotenv==1.0.1
regex==2024.9.11
requests==2.32.3
sympy==1.13.3
tiktoken==0.7.0
torch==2.2.2
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
zipp==3.20.2
